tests/requirements.txt
Whoosh
black
coverage
mypy
pylint
pylint-django
